build: remove some unneeded includes from slz.c
Przemyslaw Skibinski reported that sys/mman.h and sys/user.h break the
build on Windows. In fact these files were included in the early days
because they were needed by the test code that was migrated to zenc. It
just happens that errno, stdlib, unistd and sys/stat are not needed either.
Strictly speaking stdio is not needed but it's convenient to have it when
some printf-based debug outputs are uncommented.