optim: use bsf on x86 to compute the lower bits of the match length
On machines having a bsf instruction, it's quite convenient to be able
to immediately compute the lower bits of the match length without having
to use masks. It speeds up the compression by about 6% percent, jumping
from 402 MB/s to 432 MB/s on Silesia on a Core i7-6700k at 4.4 GHz.