medium: pre-compute the CRC32 instead of using __slz_make_crc_table()
__slz_make_crc_table() is called in the constructor on startup.
Precomputing it can improve the memory usage and/or making its
inclusion in certain projects easier.

Thus, this patch introduces a build macro PRECOMPUTE_TABLES which
then makes use of the embedded pre-computed version and removes
the corresponding initialization code. Existing programs will not
notice any difference, except the extra 4kB added to the text
section in the executable.

For now the macro is not defined.