medium: pre-compute fh_dist_table instead of using __slz_prepare_dist_table()
__slz_prepare_dist_table() is called in the constructor on startup.
Precomputing it can improve the memory usage and/or making its
inclusion in certain projects easier. It's especially true for shared
libraries where a single copy of the code will be shared among all
processes using it.

This patch relies on  the PRECOMPUTE_TABLES macro to decide whether
or not to use the pre-computed version of the table or to compute it
on startup. The ABI and API remain unchanged and the performance is
not affected. Only the text section is inflated by 128kB when used.