build: use pre-computed tables in the shared library
In a shared library it makes sense to use pre-computed tables bacause
they can save memory space when many processes use the library. Also
there is no startup time. It was verified that the performance remains
exactly the same (only off about 2.6% compared to the static library
due to -fPIC).