fix: rework and rename enqueue16() to enqueue24()
The function enqueue16() was initially designed to be called with
maximum 16 bit quantities, but it happens to be called with up to :
  - 9 bits for litterals
  - 13 bits for lengths
  - 18 bits for distances

The code doesn't resist to more than a few consecutive calls with 18
bits as it can only emit 16 bits per call. Fortunately, 13-bit calls
are interleaved with the 18 bit calls, ensuring that excess bits are
flushed in time. But the queue contains more than the max 7 bits that
are documented, and the code is fragile.

Let's rework the function to first try to dump the first 16 bits,
then the remaining 8 bits, and rename the function to match its
capabilities. Interestingly, the code is now slightly faster by
0.6%.