optim: support a 64-bit queue on 64-bit machines
The problem with a 32-bit queue is that it is not possible to keep
multiple bytes in queue to hope for a grouped write because the next
write may overflow what was kept. Eg, if the queue keeps 15 bits, a
18-bit write will cause an overflow. It is possible to avoid this by
adding preliminary checks before the queue, but this costs time.

A better solution consists in using a 64-bit queue, which has no
problem queuing 18 bits after 32 bits. But 64-bit shifts on 32-bit
machines are very expensive.

So now we use a 64-bit queue only on 64-bit machines, with 32-bit
writes at once. The queueing functions are adapted accordingly,
and since the function contains a single test, both enqueue8() and
enqueue24() can be merged.

A nice 5% extra performance improvement was achieved with this change.