zenc: fix confusing input size computation
The input size may be forced using -b but the variable names are confusing
(eg: bufsize) and this confusion has led to the fact that the buffer size
is not properly rounded up to a page size when forced. This patch fixes
this and renames the variables to "toread" to indicate how many bytes have
to be read, and "mapsize" to indicate that this size is only for mmap().