zenc: double the output buffer size
By having an output buffer double the input one, we can significantly
reduce the number of write() calls because most of the time there's
enough room left for a full input buffer after a first block was
already compressed. This is even more visible with large buffers
when dealing with pipes on input.