zenc: make the block size configurable using -1..-9
The block size used to be 32kB only for experimentation purposes but
if the utility is used to compress useful data it makes sense to
unlock this in order to achieve slightly better compression ratios.
For this we use -1, -2 and -3..-9 to respectively set the smallest
buffer, the medium one and the largest one. The difference in speed
and compression ratio is not huge, it varies for about 4% each. The
default has been set to the highest one (-3) which involves 1 MB
blocks because it's the fastest on pipes, which is the most likely
use case of this utility.