zenc: try to read larger input blocks
When reading from a pipe, there's a risk to read only small blocks,
leading to a less optimal compression ratio. This is particularly
the case when reading tar's output which defaults to 10kB blocks.
But spending too much time waiting for a sender can also force the
sender to pause on a full pipe, which is not optimal either. Here
we try to read at least a third of a buffer, hoping to be able to
compress this part before the sender finishes to fill the pipe.

A test run on a Core i5-3320M at 3.1 GHz archiving the Linux kernel 4.10
using git archive shows this :

 $ git archive --format=tar v4.10 | { nothing|gzip -n|zenc }

   command            | time(s)| size(MB)| CPU usage
 ---------------------+--------+---------+-------------------------
   git archive alone  |  4.063 |  713.1  | git at 100%
   git archive| gzip  | 23.701 |  145.2  | git at 16%, gzip at 100%
   git archive|gzip -1| 10.542 |  181.5  | git at 40%, gzip at 100%
   git archive| zenc  |  4.055 |  224.9  | git at 100%, zenc at 66%

It's interesting to note that the compression remains transparent on
this dual-core machine since it's faster to compress the data than to
produce them. When run on the same core it's possible to observe the
total work inflicted to the machine :

   command            | time(s)| size(MB)| CPU usage
 ---------------------+--------+---------+-------------------------
   git archive alone  |  4.063 |  713.1  | git at 100%
   git archive| gzip  | 26.903 |  145.2  | git at 20%, gzip at 80%
   git archive|gzip -1| 13.756 |  181.5  | git at 24%, gzip at 76%
   git archive| zenc  |  6.997 |  225.0  | git at 54%, zenc at 46%