bug: fix bug when dealing with literals 64k or larger
Nicolas Boos reported that zenc -3 produces an incorrect stream on
certain inputs like glibc-2.19.tar. It was narrowed down to an
incompressible 64kB block emitted as a literal, which is sent in two
parts (first 65535 and second 1 byte), but the source buffer is not
updated between the two iterations, thus the first byte of the buffer
is copied twice.