zenc: report an error on failure to write() to output
Some distros emit a warning when write() result is not tested. While
at first glance it seems pointless to test this for stdout, in practice
failures will often indicate a file system full condition when producing
an output file so it's better to test for this and report an error when
it happens.