slz: hash using the CRC32 instruction when available
Some ARMv8 platforms have a CRC32 instruction which provides a slightly
better hash than the one we use, in less CPU cycles. This patch makes
use of this feature when it's detected. It's typically needed to build
using -march=native on most armv8 platforms to have it. The compression
ratio slightly improves (0.1%) and the compression time is reduced by
about 1% on cortex A72.