bug: fix modulus in optimized adler32 checksum calculation for zlib encoding
The adler32 block-mode checksum calculation that was brought by commit
36aa1c0 ("optim: use a set of shifts to avoid the modulus") is bogus,
when the output value is exactly 65521 it's not set back to zero. This
means that in practice 1 out of 65536 blocks gets a bogus checksum.
With 16kB input blocks, this means one error on average every gigabyte
of input data when using zlib encoding. This encoding is used when HTTP
compression "deflate" is negotiated between a client and a server, and
may result in occasionally broken objects. Fortunately "gzip" is more
often used.

Many thanks to Pascal Vergnory for spotting this issue and reporting
it with a reproducer.