zenc: release pages after compression
Let's use madvise() when available to signal that the input file will
be sequential, and release pages by 2MB blocks after compression (this
seems to give the best trade-off of static memory usage and performance).

Now compressing a 20GB file only results in a few 10s to 100s of MB of
RAM being used.