zenc: configure the output pipe size
It's common to see zenc not being able to use 100% of the CPU just because
it's piping its output to a slow reader. A typical example is zenc|zcat
which shows 40% and 65% CPU usage respectively, but it could have even
more impact when piping data over the network.

Just setting an output buffer to be as large as the block size is enough
not to wait anymore and to compress at least at the speed accepted by the
next process. Example:

  $ time ./zenc-before /dev/shm/100m | zcat >/dev/null
  real    0m2.587s
  user    0m2.633s
  sys     0m0.049s

  $ time ./zenc-after /dev/shm/100m | zcat >/dev/null
  real    0m1.645s
  user    0m2.626s
  sys     0m0.076s

This is 36% or one second of transfer time that are saved.