slz: fix broken initialization of CRC tables
A typo was introduced in commit d8715db ("slz: do not produce the
crc32_fast table when CRC is natively supported") causing the
condition to initialize the tables to be inverted. As a result,
machines which do *not* have a hardware CRC32 instruction do not
initialize the tables. The code which was tested on haproxy did
have it, and the tests run on ARM to validate the build didn't
show any error.

Thanks to Nicolas Boos for reporting the problem!