build: fix share linking on AIX
AIX needs -Wl,-G to build shared objects instead of -Wl,-soname,blah.
The soname isn't stored into the file. Instead when a program wants
to link against a .so it may pass "-Wl,-brtl" to search for .so (thus
without any soname), or it may simply reference the .so.1 as one of
the objects to link, in which case that name will be referenced by
the target program.