zenc: use BLKGETSIZE64 on Linux to find a block device size
It's sometimes particularly convenient to compress a whole hard
disk or partition to send it over the network and save the transfer
of empty areas. Unfortunately the call to fstat() doesn't return a
block device size, so mmap() isn't used and instead read() is used.

On Linux we have an ioctl for this, so let's use it when possible.
Sometimes it seems that adjusting the block device's readahead to
~2 MB helps quite a bit.