MAJOR: reorganize libslz to split compression API code from common API code
This is a prerequisite change in order to prepare for the uncompress part
of the API. Both common and compress specific functions and helpers are
merged in slz.c and table.h. In this patch we split distinct code parts:

all common code (will be common to compress and uncompress API) is moved
to slz_common.c while compress specific code stays in slz.c.

To remain backward-compatible with previous versions, historical functions
that will become common to compression and decompression are kept with
"slz_" prefix inside slz_common.c source file, but if new common functions
were to be added, they should be prefixed with "slz_common_" instead as it
is less confusing.

Special work was carried out in table.h as it was not possible to include
the header from 2 distinct sources files due to the use of static arrays
combined with library constructors (the arrays would be available only
in the file which called the constructor), so table.h was reworked so
that it now only contains the precomputed arrays, letting the
responsibility of the recomputation to source files. It means crc
array recomputation is performed in slz_common.c while huffman dist table
array recomputation in slz.c.

No change of behavior is expected.